requirements.txt
Django>=4.2,<4.3
dj-database-url==2.1.0
psycopg[binary]>=3.1,<3.2
gunicorn>=21.2
uvicorn>=0.30.1
django-storages>=1.14
boto3>=1.34
django-unfold==0.20.5
whitenoise[brotli]==6.6.0
django-environ==0.11.2
django-ninja>=1.2.2
PyJWT>=2.9.0
django-cors-headers>=4.5.0